Add test that checks if bzPoints produces the correct number of points.
The test only checks one case right now: bzPoints(64, 2). A fake
implmentation of Brillouin zone traversal was built to make the test pass.

In order to get bzPoints to run without looping forever, it was necessary
to remove the done channel and instead have the points channel produce nil
when finished. Not sure why this was needed.